+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Non json-serializable data in DataStorage instances are now saved to a single pickle file
 - Prevent NameError being raised if an exception was raised inside nested profiling without the use of `with TT.profile`
 - Added `string` method to `HardwareInfo` for easy formatting
+- Removed `no_grad` decorator
 
 ### Bug fixes